LE_Kotlin: just renamed "logic" -> "gameLogic"
