+++ Attt_LocalEngine_Kotlin/src/main/kotlin/gameLogic/GameRules.kt
@@ -1,4 +1,4 @@
-package logic
+package gameLogic
 
 import constants.MAX_WINNING_LINE_LENGTH
 import constants.MIN_WINNING_LINE_LENGTH
