+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept2D/NearestAreaScanWith2D.kt
@@ -1,7 +1,7 @@
 package geometry.concept2D
 
 import geometry.Coordinates
-import logic.GameField
+import gameLogic.GameField
 import geometry.OneMoveProcessing
 import players.PlayerProvider
 import utilities.Log
