+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept3D/NearestAreaScanWith3D.kt
@@ -1,7 +1,7 @@
 package geometry.concept3D
 
 import geometry.Coordinates
-import logic.GameField
+import gameLogic.GameField
 import geometry.OneMoveProcessing
 import players.PlayerProvider
 import utilities.Log
