+++ Attt_LocalEngine_Kotlin/src/main/kotlin/publicApi/AtttGame.kt
@@ -3,7 +3,7 @@ package publicApi
 import constants.MIN_GAME_FIELD_SIDE_SIZE
 import constants.MIN_NUMBER_OF_PLAYERS
 import constants.MIN_WINNING_LINE_LENGTH
-import logic.GameSession
+import gameLogic.GameSession
 
 interface AtttGame {
